test(themes): address review nits
